src/main/java/io/zipcoder/polymorphism/Zebra.java
package io.zipcoder.polymorphism;

public class Zebra extends Pet {
private String type;
private String name;
    public Integer NumberOfZebras;
    public Zebra() {
    }

    public Zebra(String type, String name) {
        this.type = type;
        this.name = name;
    }

    public void setName(String name) { this.name = name; }

    public String getName() {
        return name;
    }

    public String speak() {
        return "Woo";
    }

    public Integer getNumberOfZebras(){
        return NumberOfZebras;
    }
}

